requirements.txt
Flask==2.0.1
line-bot-sdk==3.11.0
openai>=1.0.0
Werkzeug==2.0.3
